// src/app/core/typings/gobs1.typings.ts
export class Gobs1Typings {
    CARAC: string = null;
    CT: number = 0;
    DLA: Date = new Date();
    Etat: string = null;
    Id: number = 0;
    IdMeute: number = 0;
	IdClan: number = 0;
    N: number = 0;
    Niveau: number = 0;
    Nom: string = null;
    NomMeute: string = null;
    PA: number = 0;
    PI: number = 0;
    PV: number = 0;
    PX: number = 0;
    PXPerso: number = 0;
    Race: string = null;
    Tribu: string = null;
    X: number = 0;
    Y: number = 0;
    Z: number = 0;
}